requirements.txt
langchain
ipykernel
python-dotenv
streamlit
langchain_community
langchain_core
langchain-groq